fineeeee

+++ fine/elimina_utente.php
@@ -1,11 +1,10 @@
-<?php
-    
-    require "./funzioni.php";
-    $conn = db_conn();
-    
+<?php    
+require "./funzioni.php";
+$conn = db_conn();
+
+if(isset($_SESSION['adminLog']) && $_SESSION['adminLog']=='daje'){
     $dropThis = $_POST['dropThis'];
-    
-    
+       
     $sql = "DELETE FROM spesa WHERE spesa.utente = '$dropThis'";
     if((mysqli_query($conn, $sql))){
         $sql = "DELETE FROM utente WHERE utente.username = '$dropThis'";
@@ -15,5 +14,7 @@
     }else{
         header("Location: view.php?error=qualcosa è andato storto");
     }
-        
+}else{
+    header("Location: index.php");
+}       
 ?>
